+++ rspack.js
@@ -1,1 +1,4 @@
-module.exports = require('@rspack/core')
+import rspack from '@rspack/core'
+
+export * from '@rspack/core'
+export default rspack