+++ test/fixtures/pkg-with-config/jetpack.config.mjs
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   port: 1234,
   entry: './app/client',
   title: 'testing',