+++ test/fixtures/pkg-with-esm/jetpack.config.mjs
@@ -1,11 +1,7 @@
-const rspackConfigForTests = require('../../helpers/rspackConfigForTests')
+import rspackConfigForTests from '../../helpers/rspackConfigForTests.js'
 
-module.exports = {
+export default {
   minify: false,
-  target: {
-    modern: true,
-    legacy: true
-  },
   rspack(config) {
     delete config.optimization.splitChunks
     delete config.optimization.runtimeChunk