+++ test/fixtures/pkg-with-lightningcss/jetpack.config.mjs
@@ -1,0 +1,3 @@
+export default {
+  minify: false
+}